# src/doc/swagger.yaml
openapi: 3.0.4
info:
  title: Blog API
  version: 2.0.0
  description: "API documentation for personal blogging platform"
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
paths:
...
  /posts/{postId}/comments/{commentId}/replies:
    get:
      summary: Get replies to a comment
      operationId: getCommentReplies
      description: |
        Fetches replies for a specific comment using cursor-based pagination.
        Replies are returned from oldest to newest. If a `cursor` is provided, it will fetch replies starting from that ID.

      tags: [Replies]
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"

        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 20
          description: Max number of replies to return (defaults to 20, max 100)

        - $ref: "#/components/parameters/CursorParam"

      responses:
        200:
          description: List of replies with optional pagination cursor
          content:
            application/json:
              schema:
                type: object
                properties:
                  cursor:
                    type: string
                    nullable: true
                    description: The ObjectId of the next reply, if more results exist
                    example: "66aee40d1540c8321183c5f0"
                  replies:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          example: "66aee40d1540c832e183c5f4"
                        replyToUsername:
                          type: string
                        body:
                          type: string
                        likes:
                          type: integer
                        author:
                          type: object
                          properties:
                            _id:
                              type: string
                              example: "66aee40d1540c8321183c5f03"
                            username:
                              type: string
                            profileImage:
                              type: string
                              nullable: true
                        createdAt:
                          type: string
                          format: date-time
        400:
          description: Bad request, invalid post id, comment id or cursor parameter.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "cursor is not valid"
        404:
          $ref: "#/components/responses/PostOrCommentNotFound"
        500:
          $ref: "#/components/responses/ServerError"

    post:
      summary: Create a reply to a comment
      operationId: createReply
      description: |
        Creates a reply on a comment within a post. Only authenticated users can reply.
      tags: [Replies]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
                - replyToUsername
              properties:
                body:
                  type: string
                  description: The content of the reply
                  example: "I totally agree with your comment!"
                replyToUsername:
                  type: string
                  description: Username of the person being replied to
                  example: "ahmed_dev"

      responses:
        201:
          $ref: "#/components/responses/Created"
        400:
          description: Validation errors in the request body or path
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    properties:
                      postId:
                        type: string
                        nullable: true
                        example: "Invalid postId"
                      commentId:
                        type: string
                        nullable: true
                        example: "Invalid commentId"
                      replyToUsername:
                        type: string
                        nullable: true
                        example: "Username of the user being replied to is required"
                      body:
                        type: string
                        nullable: true
                        example: "Body must not be empty"
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: "#/components/responses/PostOrCommentNotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /posts/{postId}/comments/{commentId}/replies/{replyId}:
    patch:
      summary: Update a reply
      operationId: updateReply
      description: |
        Updates the body of a specific reply. Only the reply's author can perform this action.
      tags: [Replies]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
        - $ref: "#/components/parameters/ReplyIdParam"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  example: "Updated reply content."

      responses:
        200:
          $ref: "#/components/responses/Updated"
        400:
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    properties:
                      postId:
                        type: string
                        nullable: true
                        example: "Invalid postId"
                      commentId:
                        type: string
                        nullable: true
                        example: "Invalid commentId"
                      replyId:
                        type: string
                        nullable: true
                        example: "Invalid replyId"
                      body:
                        type: string
                        nullable: true
                        example: "Body must not be empty"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/PostOrCommentOrReplyNotFound"
        500:
          $ref: "#/components/responses/ServerError"

    delete:
      summary: Delete a reply
      operationId: deleteReply
      description: |
        Deletes a specific reply to a comment.
        Only the reply's author, an admin, or a moderator can perform this operation.
      tags: [Replies]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
        - $ref: "#/components/parameters/ReplyIdParam"

      responses:
        200:
          $ref: "#/components/responses/Deleted"
        400:
          $ref: "#/components/responses/InvalidResourceId"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/PostOrCommentOrReplyNotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /posts/{postId}/comments/{commentId}/replies/{replyId}/likes:
    post:
      summary: Like a reply
      operationId: likeReply
      description: |
        Likes a specific reply within a comment. If already liked by the user, the server responds with a success message but takes no further action.
      tags: [Replies]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
        - $ref: "#/components/parameters/ReplyIdParam"

      responses:
        200:
          $ref: "#/components/responses/Success"
        400:
          $ref: "#/components/responses/InvalidResourceId"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/PostOrCommentOrReplyNotFound"
        500:
          $ref: "#/components/responses/ServerError"

    delete:
      summary: Unlike a reply
      operationId: unlikeReply
      description: |
        Removes the like from a specific reply made by the authenticated user.
        Does nothing if the user hasn’t liked the reply before.
      tags: [Replies]
      security:
        - JWTAuth: []

      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
        - $ref: "#/components/parameters/ReplyIdParam"

      responses:
        200:
          $ref: "#/components/responses/Success"
        400:
          $ref: "#/components/responses/InvalidResourceId"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/PostOrCommentOrReplyNotFound"
        500:
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Use the JWT token obtained from login or registration to access protected routes.
  parameters:
    PostIdParam:
      in: path
      name: postId
      required: true
      schema:
        type: string
        format: objectId
        example: "64f0341f3df05e12a3c3a401"
      description: ID of the post
    CommentIdParam:
      in: path
      name: commentId
      required: true
      schema:
        type: string
        format: objectId
        example: "64f0341f3df05212c3c3b402"
      description: ID of the comment
    ReplyIdParam:
      in: path
      name: replyId
      required: true
      schema:
        type: string
        example: "64f0341f3df05212c3c3b42e"
      description: ID of the reply
    CursorParam:
      in: query
      name: cursor
      schema:
        type: string
        nullable: true
        example: 64f0341f3df05e12a3c3a401
      description: Cursor to the next page (null if no more)
...
  responses:
...
    Success:
      description: Success, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Success
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Created:
      description: Created, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Created
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Updated:
      description: Updated, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Updated
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Deleted:
      description: Deleted, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Deleted
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
...
    InvalidResourceId:
      description: Bad request due to invalid post, comment, or reply ID.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Invalid comment reply id"
...
    PostOrCommentNotFound:
      description: Post or comment not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Comment not found"
    PostOrCommentOrReplyNotFound:
      description: Post or comment or reply not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Post not found"
    Unauthorized:
      description: Unauthorized – user not authenticated or token is invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    Forbidden:
      description: Forbidden – user lacks permission (read description to know what roles are authorized)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Forbidden
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error message
...